# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OfferDTO": {
        "additionalProperties": false,
        "description": "One trusted server-side offer (clients pick offer_key only).",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "offer_key": {
            "title": "Offer Key",
            "type": "string"
          },
          "points_amount": {
            "title": "Points Amount",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "offer_key",
          "version",
          "description",
          "amount",
          "currency",
          "points_amount"
        ],
        "title": "OfferDTO",
        "type": "object"
      },
      "OfferListDTO": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OfferDTO"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "OfferListDTO",
        "type": "object"
      },
...
      "PurchaseInput": {
        "additionalProperties": false,
        "properties": {
          "offer_key": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Offer Key",
            "type": "string"
          }
        },
        "required": [
          "offer_key"
        ],
        "title": "PurchaseInput",
        "type": "object"
      },
      "PurchaseOrderDTO": {
        "additionalProperties": false,
        "description": "Purchase workflow checkout view returned at start (and on replay).",
        "properties": {
          "checkout_url": {
            "title": "Checkout Url",
            "type": "string"
          },
          "order_reference": {
            "title": "Order Reference",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "order_reference",
          "checkout_url",
          "state"
        ],
        "title": "PurchaseOrderDTO",
        "type": "object"
      },
...
      "RefundDTO": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "refund_ref": {
            "title": "Refund Ref",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_id",
          "refund_ref",
          "amount",
          "currency",
          "state",
          "reason"
        ],
        "title": "RefundDTO",
        "type": "object"
      },
...
      "RequestRefundInput": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "reason",
          "idempotency_key"
        ],
        "title": "RequestRefundInput",
        "type": "object"
      },
...
      "SettingGroupDTO": {
        "additionalProperties": false,
        "properties": {
          "group_key": {
            "title": "Group Key",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "group_key",
          "schema_version",
          "version",
          "values"
        ],
        "title": "SettingGroupDTO",
        "type": "object"
      },
...
      "UpdateSettingGroupInput": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "title": "UpdateSettingGroupInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WebhookReceiptDTO": {
        "additionalProperties": false,
        "properties": {
          "duplicate": {
            "default": false,
            "title": "Duplicate",
            "type": "boolean"
          },
          "received": {
            "title": "Received",
            "type": "boolean"
          }
        },
        "required": [
          "received"
        ],
        "title": "WebhookReceiptDTO",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "aiya-cms (cms)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/payments/orders/{order_id}/refund": {
      "post": {
        "operationId": "request_refund_api_v1_admin_payments_orders__order_id__refund_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRefundInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Refund",
        "tags": [
          "admin",
          "admin-payments"
        ]
      }
    },
...
    "/api/v1/admin/settings/groups": {
      "get": {
        "operationId": "list_groups_api_v1_admin_settings_groups_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SettingGroupDTO"
                  },
                  "title": "Response List Groups Api V1 Admin Settings Groups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Groups",
        "tags": [
          "admin",
          "admin-settings"
        ]
      }
    },
    "/api/v1/admin/settings/groups/{group_key}": {
      "get": {
        "operationId": "get_group_api_v1_admin_settings_groups__group_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_key",
            "required": true,
            "schema": {
              "title": "Group Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingGroupDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Group",
        "tags": [
          "admin",
          "admin-settings"
        ]
      },
      "put": {
        "operationId": "update_group_api_v1_admin_settings_groups__group_key__put",
        "parameters": [
          {
            "in": "path",
            "name": "group_key",
            "required": true,
            "schema": {
              "title": "Group Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingGroupInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingGroupDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Group",
        "tags": [
          "admin",
          "admin-settings"
        ]
      }
    },
    "/api/v1/admin/settings/groups/{group_key}/reset": {
      "post": {
        "operationId": "reset_group_api_v1_admin_settings_groups__group_key__reset_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_key",
            "required": true,
            "schema": {
              "title": "Group Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingGroupDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Group",
        "tags": [
          "admin",
          "admin-settings"
        ]
      }
    },
...
    "/api/v1/point-purchase/offers": {
      "get": {
        "operationId": "list_offers_api_v1_point_purchase_offers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferListDTO"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Offers",
        "tags": [
          "point-purchase"
        ]
      }
    },
    "/api/v1/point-purchase/orders": {
      "post": {
        "operationId": "start_purchase_api_v1_point_purchase_orders_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrderDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Purchase",
        "tags": [
          "point-purchase"
        ]
      }
    },
...
    "/api/v1/webhooks/payments/{provider_key}": {
      "post": {
        "operationId": "payment_webhook_api_v1_webhooks_payments__provider_key__post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_key",
            "required": true,
            "schema": {
              "title": "Provider Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookReceiptDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Payment Webhook",
        "tags": [
          "webhooks"
        ]
      }
    },
...
  },
...
}